tests: move remote storage related wrappers out of neon_fixtures.py
