+++ test_runner/regress/test_timeline_delete.py
@@ -13,9 +13,6 @@ from fixtures.neon_fixtures import (
     NeonEnv,
     NeonEnvBuilder,
     PgBin,
-    RemoteStorageKind,
-    S3Storage,
-    available_remote_storages,
     last_flush_lsn_upload,
     wait_for_last_flush_lsn,
 )
@@ -27,6 +24,11 @@ from fixtures.pageserver.utils import (
     wait_timeline_detail_404,
     wait_until_tenant_active,
     wait_until_timeline_state,
+)
+from fixtures.remote_storage import (
+    RemoteStorageKind,
+    S3Storage,
+    available_remote_storages,
 )
 from fixtures.types import Lsn, TenantId, TimelineId
 from fixtures.utils import query_scalar, wait_until
